Rearranging stat macros
