+++ src/module/macros.ts
@@ -26,7 +26,7 @@ import {
   TagInstance,
 } from "machine-mind";
 import { FoundryFlagData, FoundryReg } from "./mm-util/foundry-reg";
-import { is_ref, resolve_dotpath } from "./helpers/commons";
+import { is_ref } from "./helpers/commons";
 import { checkForHit } from "./helpers/automation/targeting";
 import type { AccDiffData, AccDiffDataSerialized, RollModifier } from "./helpers/acc_diff";
 import { is_limited, is_overkill } from "machine-mind/dist/funcs";
@@ -36,7 +36,6 @@ import { getAutomationOptions } from "./settings";
 
 import { getMacroSpeaker, encodeMacroData, encodedMacroWhitelist, ownedItemFromString } from "./macros/util"
 import { renderMacroTemplate } from "./macros/render"
-import { rollTechMacro } from "./macros/tech"
 import { prepareTextMacro, rollTextMacro } from "./macros/text"
 
 export { encodeMacroData, runEncodedMacro } from "./macros/util"
@@ -45,6 +44,7 @@ export { prepareActivationMacro } from "./macros/activation"
 export { prepareItemMacro } from "./macros/item"
 export { prepareOverchargeMacro } from "./macros/overcharge"
 export { stabilizeMacro } from "./macros/stabilize"
+export { prepareStatMacro } from "./macros/stat"
 export { prepareTechMacro } from "./macros/tech"
 export { prepareTextMacro } from "./macros/text"
 
@@ -239,74 +239,6 @@ export function attackRolls(bonus: number, accdiff: AccDiffData): AttackRolls {
       };
     }),
   };
-}
-
-export async function prepareStatMacro(a: string, statKey: string, rerollData?: AccDiffDataSerialized) {
-  // Determine which Actor to speak as
-  let actor = getMacroSpeaker(a);
-  if (!actor) return;
-
-  const statPath = statKey.split(".");
-
-  let mm_ent = await actor.data.data.derived.mm_promise;
-
-  let bonus: number = resolve_dotpath(mm_ent, statKey.substr(3));
-
-  let mData: LancerStatMacroData = {
-    title: statPath[statPath.length - 1].toUpperCase(),
-    bonus: bonus,
-  };
-  if (mData.title === "TECHATTACK") {
-    let partialMacroData = {
-      title: "Reroll stat macro",
-      fn: "prepareStatMacro",
-      args: [a, statKey],
-    };
-    rollTechMacro(
-      actor,
-      { acc: 0, action: "Quick", t_atk: bonus, effect: "", tags: [], title: "" },
-      partialMacroData,
-      rerollData
-    );
-  } else {
-    rollStatMacro(actor, mData).then();
-  }
-}
-
-// Rollers
-
-export async function rollStatMacro(actor: LancerActor, data: LancerStatMacroData) {
-  if (!actor) return Promise.resolve();
-
-  // Get accuracy/difficulty with a prompt
-  let { AccDiffData } = await import("./helpers/acc_diff");
-  let initialData = AccDiffData.fromParams(actor, undefined, data.title);
-
-  let promptedData;
-  try {
-    let { open } = await import("./helpers/slidinghud");
-    promptedData = await open("hase", initialData);
-  } catch (_e) {
-    return;
-  }
-
-  let acc: number = promptedData.base.total;
-
-  // Do the roll
-  let acc_str = acc != 0 ? ` + ${acc}d6kh1` : "";
-  let roll = await new Roll(`1d20+${data.bonus || 0}${acc_str}`).evaluate({ async: true });
-
-  const roll_tt = await roll.getTooltip();
-
-  // Construct the template
-  const templateData = {
-    title: data.title,
-    roll: roll,
-    roll_tooltip: roll_tt,
-    effect: data.effect ? data.effect : null,
-  };
-  const template = `systems/${game.system.id}/templates/chat/stat-roll-card.hbs`;
-  return renderMacroTemplate(actor, template, templateData);
 }
 
 type AttackMacroOptions = {
